move comparisons scripts to separate folder
