+++ splink/comparison/comparison_level_library.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from .comparison_level import ComparisonLevel
 from .comparison_level_sql import great_circle_distance_km_sql
-from .input_column import InputColumn
+from ..input_column import InputColumn
 
 
 class NullLevelBase(ComparisonLevel):
