+++ splink/comparison/comparison_library.py
@@ -7,7 +7,7 @@ from .comparison_library_utils import (
     distance_threshold_comparison_levels,
     distance_threshold_description,
 )
-from .misc import ensure_is_iterable
+from ..misc import ensure_is_iterable
 
 
 class ExactMatchBase(Comparison):
