+++ splink/spark/spark_comparison_template_library.py
@@ -1,4 +1,4 @@
-from ..comparison_template_library import (
+from ..comparison.comparison_template_library import (
     DateComparisonBase,
     NameComparisonBase,
 )
